feat(delete): Added question if you really want to delete a record

+++ client/src/preceptor/handlers/HandlersPreceptor.js
@@ -1,6 +1,7 @@
 import iziToast from "izitoast"
 import 'izitoast/dist/css/iziToast.min.css';
 import { fetchPreceptor } from "../services/PreceptorService";
+import Swal from 'sweetalert2';
 
 // Función auxiliar para manejar un registro exitoso
 export const handleRegistroSuccess = (data, reset, navigate, createPreceptor) => {
@@ -37,24 +38,39 @@ export const handleEditSuccess = (data, navigate, updateStudent) => {
     navigate("/IPF/students/")
 }
 
-export const handleDeletePreceptor = async(id, deletePreceptor) => {
-  try {
-    const data = await fetchPreceptor(`users/${id}`, "DELETE", null)
+export const handleDeletePreceptor = async (id, deletePreceptor) => {
+  // Muestra la confirmación al usuario
+  const result = await Swal.fire({
+    title: '¿Estás seguro?',
+    text: "No podrás revertir esto",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Sí, eliminar',
+    cancelButtonText: 'Cancelar',
+  });
 
-    deletePreceptor(id)
-    // Notifica al usuario
-    iziToast.success({
-      title: "Éxito",
-      message: data.message,
-      position: "topRight",
-    });
-    
-  } catch (error) {
-    iziToast.error({
+  // Verifica si el usuario confirmó la eliminación
+  if (result.isConfirmed) {
+    try {
+      const data = await fetchPreceptor(`users/${id}`, "DELETE", null);
+
+      deletePreceptor(id);
+
+      // Notifica al usuario de éxito
+      iziToast.success({
+        title: "Éxito",
+        message: data.message,
+        position: "topRight",
+      });
+    } catch (error) {
+      iziToast.error({
         title: "Error",
         message: "Error interno en el servidor",
         position: "topRight",
-    });
-    console.error("Error al eliminar el preceptor", error)
+      });
+      console.error("Error al eliminar el preceptor", error);
+    }
   }
-}
+};
